requirements-dev.txt
# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-flask==1.3.0
pytest-asyncio==0.21.1

# Code quality
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
bandit==1.7.5

# Documentation
sphinx==7.0.1
sphinx-rtd-theme==1.2.2
sphinx-autodoc-typehints==1.23.0

# Development tools
pre-commit==3.3.3
ipython==8.14.0
python-dotenv==1.0.0

# Test dependencies
coverage==7.2.7
factory-boy==3.3.0
Faker==19.3.0
responses==0.23.1

# Linting
pylint==3.0.0
dlint==0.14.0

# Type checking
types-requests==2.31.0.1
types-python-dateutil==2.8.19.13

# Security
safety==2.3.5

# Debugging
pdbpp==0.10.3

# Formatting
pre-commit-hooks==4.4.0
yapf==0.40.1

# Container tools
docker==6.1.3
docker-compose==1.29.2

# API testing
httpx==0.24.1
pytest-httpx==0.24.1

# Database testing
mongomock==4.1.2

# Test reporting
pytest-html==4.0.2
pytest-xdist==3.3.1

# Code coverage
codecov==2.1.13
